Use fixed number format for error messages in `torch.testing`
By default numbers are formatted with the "General format" (`:g`). This might lead to diverging formats of two related numbers if they have different magnitudes:

```python
abs_diff = 1e-4
atol = 1e-5
print(f"Absolute difference: {abs_diff} (up to {atol} allowed)")
```

```
Absolute difference: 0.0001 (up to 1e-05 allowed)
```

From this message it is hard to parse at a glance if the value is inside the specified tolerance or not.

To overcome this I've added `:.2e` as a format since this notation is usually used to define the tolerances.

ghstack-source-id: 2ae24dc9833d5b846a89825277094b9503f84582
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57134