[pytorch] gate TLS RecordFunctionCallbacks for Android with libgnustl
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57580

The thread_local on Android with old libgnustl was broken. It can run
thread_local object's destructor after it frees the memory. So it's unsafe
to have thread_local objects with non-trivial destructor.

This diff removed sorted_tls_callbacks_ from the RecordFunctionTLS
struct, which is a std::vector with non-trivial destructor. This should
fix some crashes we saw in production.
ghstack-source-id: 128165096

Differential Revision: [D28208027](https://our.internmc.facebook.com/intern/diff/D28208027/)