introduce TestingErrorMeta for internal use
In #54780 we opted to return `Optional[Exception]` from all internal
helper functions. Since then multiple PRs added functionality that needs
to amend the error message. For this we recreate the error

https://github.com/pytorch/pytorch/blob/09a1b1cf87fb8724e644693f2c00318ed2637095/torch/testing/_asserts.py#L417-L430

To untangle this a little, this PR introduces the `_TestingErrorMeta`,
which carries the exception type and the message. The idiom

```python
exc = check_foo():
if exc:
    return exc
```

is still valid although `exc` should be renamed to `error_meta` to
reflect the new nature. In the top-level functions
`assert_(equal|close)`

```python
exc = check_foo():
if exc:
    raise exc
```

changes to

```python
error_meta = check_foo():
if error_meta:
    raise error_meta.to_error()
```

ghstack-source-id: 5c10b199d11a1611ab93ff4acd0acb5020c8c409
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58917