[deploy] Make load_library a no-op inside a package
**Summary**
This commit makes load_library calls no-ops inside packages run with
deploy. Libraries containing custom C++ operators and classes are statically linked in C++
and don't need to be loaded. This commit takes advantage of the fact that sys.executable is
set to torch_deploy in deploy and uses that to exit early in load_library if
the program is running inside deploy.

**Test Plan**
This commit adds a test to `generate_examples`/`test_deploy` that
packages and runs a function that calls `load_library`. The library
doesn't exist, but that's okay because the function should be a no-op
anyway.

ghstack-source-id: 1a1b6fd34ca8b7ac7d30d9cd562d5332412ef2d9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58933