[reland] Make TP agent use streams from Future when sending response
Reland of https://github.com/pytorch/pytorch/pull/58428

Until now, the TP agent expected the output of a remote function to be on the same streams as the inputs. In other words, it used the lazy stream context of the inputs to synchronize the output tensors. This was true in the most common case of a synchronous remote function. However it wasn't true for async functions, for fetching RRefs, ... The more generic way is to use the CUDA events held by the Future to perform this synchronization. (These events may be on the input streams, or they may not be!).

Differential Revision: [D28623885](https://our.internmc.facebook.com/intern/diff/D28623885/)

ghstack-source-id: 130202842
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59212