Recognize transposed dense tensors as a form of partial overlap
Fixes #48401

`assert_no_overlap` currently currently has a false-negative where it recognizes
the transpose of a contiguous tensor as fully overlapping. This happens because
the memory regions do fully overlap, but of course the strides are different so
the actual elements don't all overlap.

This goes slightly in the other direction, where requiring strides to exactly
match which gives false-positives for some unusual situations, e.g.
```
torch.add(a, a, out=a.view([1] + a.shape)
```
Or replacing strides of length-1 dimensions, etc. However, I think these are
sufficiently obscure that it's okay to error and the common cases like
inplace operations still work as before.

ghstack-source-id: cc63ef45ce8cba8da1020aeb296dc1f1f86e7a94
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59014